requirements.txt
gevent==1.4.0
Werkzeug==0.15.2
Flask==1.0.2
python-dotenv==0.10.3
amadeus==3.1.0